# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "义乌商贸库存管家",
    "version": "0.3.0"
  },
  "paths": {
    "/api/v1/session": {
      "get": {
        "summary": "Session",
        "operationId": "session_api_v1_session_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Login",
        "operationId": "login_api_v1_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Logout",
        "operationId": "logout_api_v1_session_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/api/v1/inventory": {
      "get": {
        "summary": "Inventory",
        "operationId": "inventory_api_v1_inventory_get",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Warehouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ledger": {
      "get": {
        "summary": "History",
        "operationId": "history_api_v1_ledger_get",
        "parameters": [
          {
            "name": "warehouse_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Warehouse Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/receipts/confirm": {
      "post": {
        "summary": "Receive",
        "operationId": "receive_api_v1_receipts_confirm_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 120,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Receipt"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sales/confirm": {
      "post": {
        "summary": "Sell",
        "operationId": "sell_api_v1_sales_confirm_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 120,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Sale"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/skus": {
      "post": {
        "summary": "New Sku",
        "operationId": "new_sku_api_v1_skus_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 120,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SKU"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/warehouses": {
      "post": {
        "summary": "New Warehouse",
        "operationId": "new_warehouse_api_v1_warehouses_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 120,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Warehouse"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Login": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Login"
      },
...
      "Receipt": {
        "properties": {
          "warehouse_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Warehouse Id"
          },
          "sku_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Sku Id"
          },
          "expected_version": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "source": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Source"
          },
          "note": {
            "type": "string",
            "maxLength": 500,
            "title": "Note",
            "default": ""
          },
          "g": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "G"
          },
          "q": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Q",
            "default": 0
          },
          "d": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "D",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "warehouse_id",
          "sku_id",
          "expected_version",
          "source",
          "g"
        ],
        "title": "Receipt"
      },
...
      "SKU": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Code"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "spec": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Spec"
          },
          "unit": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Unit",
            "default": "件"
          },
          "barcode": {
            "type": "string",
            "maxLength": 120,
            "title": "Barcode",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "name",
          "spec"
        ],
        "title": "SKU"
      },
      "Sale": {
        "properties": {
          "warehouse_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Warehouse Id"
          },
          "sku_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Sku Id"
          },
          "expected_version": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "source": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Source"
          },
          "note": {
            "type": "string",
            "maxLength": 500,
            "title": "Note",
            "default": ""
          },
          "quantity": {
            "type": "integer",
            "maximum": 1000000.0,
            "minimum": 1.0,
            "title": "Quantity"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "warehouse_id",
          "sku_id",
          "expected_version",
          "source",
          "quantity"
        ],
        "title": "Sale"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Warehouse": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Code"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "country": {
            "type": "string",
            "pattern": "^[A-Z]{2}$",
            "title": "Country",
            "default": "CN"
          },
          "timezone": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Timezone",
            "default": "Asia/Shanghai"
          },
          "authority": {
            "type": "string",
            "enum": [
              "local",
              "manual"
            ],
            "title": "Authority",
            "default": "local"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "Warehouse"
      }
    }
  }
}